+++ src/client.tsx
@@ -173,9 +173,14 @@ export function useZodForm<
   );
 
   // Create the React Router fetcher for this form
-  const { data, Form: FetcherForm, load, state, submit } = (
+  let { data, Form: FetcherForm, load, state, submit } = (
     useFetcher()
   );
+
+  // Use action data if fetcher data is not available
+  if (!data && actionData?.intent === intent) {
+    data = actionData;
+  }
 
   /**
    * Validate the form data against the schema
@@ -364,15 +369,7 @@ export function useZodForm<
 
       // Get the default value from the current action data
       const defaultValue = path.pickFrom(
-        (
-          form?.data?.validation?.data?.intent === intent
-            ? form.data.validation.data
-            : undefined
-        ) || (
-          actionData?.validation?.data?.intent === intent
-            ? actionData.validation.data
-            : undefined
-        )
+        form?.data?.validation?.data
       );
 
       // If the default value exists