+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed a bug where action data could potentially pollute other forms
+- Fixed a bug where action data was not loaded correctly during SSR
 - Fixed a bug with regular expression HTML attribute formatting
 - Fixed a bug where validation attributes were not added for optional types